chore: Move OnBoarding Directory Path
