+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
 
 
         <activity
-            android:name=".onboarding.OnboardingActivity"
+            android:name=".feature.onboarding.OnboardingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
