+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/OnboardingViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class OnboardingViewModel:BaseViewModel() {
+}
