+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/first/OnboardingFirstFragment.kt
@@ -1,0 +1,15 @@
+package com.b1nd.alimo.feature.onboarding.first
+
+import androidx.fragment.app.viewModels
+import com.b1nd.alimo.R
+import com.b1nd.alimo.base.BaseFragment
+import com.b1nd.alimo.databinding.FragmentOnboardingFirstBinding
+
+class OnboardingFirstFragment:BaseFragment<FragmentOnboardingFirstBinding, com.b1nd.alimo.feature.onboarding.first.OnboardingFirstViewModel>(
+    R.layout.fragment_onboarding_first) {
+    override val viewModel: com.b1nd.alimo.feature.onboarding.first.OnboardingFirstViewModel by viewModels()
+
+    override fun initView() {
+//        TODO("Not yet implemented")
+    }
+}
