+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/first/OnboardingFirstViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.first
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class OnboardingFirstViewModel:BaseViewModel() {
+}
