+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/join/first/ParentJoinFirstViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.join.first
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentJoinFirstViewModel:BaseViewModel() {
+}
