+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/join/second/ParentJoinSecondViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.join.second
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentJoinSecondViewModel:BaseViewModel() {
+}
