+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/join/third/ParentJoinThirdFragment.kt
@@ -1,0 +1,16 @@
+package com.b1nd.alimo.feature.onboarding.parent.join.third
+
+import androidx.fragment.app.viewModels
+import com.b1nd.alimo.R
+import com.b1nd.alimo.base.BaseFragment
+import com.b1nd.alimo.databinding.FragmentParentJoinThirdBinding
+
+class ParentJoinThirdFragment:BaseFragment<FragmentParentJoinThirdBinding, com.b1nd.alimo.feature.onboarding.parent.join.third.ParentJoinThirdViewModel>(
+    R.layout.fragment_parent_join_third
+) {
+    override val viewModel: com.b1nd.alimo.feature.onboarding.parent.join.third.ParentJoinThirdViewModel by viewModels()
+
+    override fun initView() {
+//        TODO("Not yet implemented")
+    }
+}
