+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/join/third/ParentJoinThirdViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.join.third
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentJoinThirdViewModel:BaseViewModel() {
+}
