+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/login/first/ParentLoginFirstViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.login.first
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentLoginFirstViewModel:BaseViewModel() {
+}
