+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/pw/first/ParentFindPWFirstViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.pw.first
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentFindPWFirstViewModel:BaseViewModel() {
+}
