+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/parent/pw/second/ParentFindPWSecondViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.parent.pw.second
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class ParentFindPWSecondViewModel:BaseViewModel() {
+}
