+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/second/OnboardingSecondFragment.kt
@@ -1,0 +1,16 @@
+package com.b1nd.alimo.feature.onboarding.second
+
+import androidx.fragment.app.viewModels
+import com.b1nd.alimo.R
+
+import com.b1nd.alimo.base.BaseFragment
+import com.b1nd.alimo.databinding.FragmentOnboardingSecondBinding
+
+class OnboardingSecondFragment:BaseFragment<FragmentOnboardingSecondBinding, com.b1nd.alimo.feature.onboarding.second.OnboardingSecondViewModel>(
+    R.layout.fragment_onboarding_second) {
+    override val viewModel: com.b1nd.alimo.feature.onboarding.second.OnboardingSecondViewModel by viewModels()
+
+    override fun initView() {
+//        TODO("Not yet implemented")
+    }
+}
