+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/second/OnboardingSecondViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.second
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class OnboardingSecondViewModel:BaseViewModel() {
+}
