+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/student/first/StudentLoginViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.student.first
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class StudentLoginViewModel:BaseViewModel() {
+}
