+++ app/src/main/java/com/b1nd/alimo/feature/onboarding/third/OnboardingThirdViewModel.kt
@@ -1,0 +1,6 @@
+package com.b1nd.alimo.feature.onboarding.third
+
+import com.b1nd.alimo.base.BaseViewModel
+
+class OnboardingThirdViewModel:BaseViewModel() {
+}
