--- app/src/main/java/com/b1nd/alimo/onboarding/OnboardingViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class OnboardingViewModel:BaseViewModel() {
-}
