--- app/src/main/java/com/b1nd/alimo/onboarding/first/OnboardingFirstViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.first
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class OnboardingFirstViewModel:BaseViewModel() {
-}
