--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/first/ParentJoinFirstFragment.kt
@@ -1,16 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.first
-
-import androidx.fragment.app.viewModels
-import com.b1nd.alimo.R
-import com.b1nd.alimo.base.BaseFragment
-import com.b1nd.alimo.databinding.FragmentParentJoinFirstBinding
-
-class ParentJoinFirstFragment:BaseFragment<FragmentParentJoinFirstBinding, ParentJoinFirstViewModel>(
-    R.layout.fragment_parent_join_first
-) {
-    override val viewModel: ParentJoinFirstViewModel by viewModels()
-
-    override fun initView() {
-//        TODO("Not yet implemented")
-    }
-}
