--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/first/ParentJoinFirstViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.first
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class ParentJoinFirstViewModel:BaseViewModel() {
-}
