--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/second/ParentJoinSecondFragment.kt
@@ -1,16 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.second
-
-import androidx.fragment.app.viewModels
-import com.b1nd.alimo.R
-import com.b1nd.alimo.base.BaseFragment
-import com.b1nd.alimo.databinding.FragmentParentJoinSecondBinding
-
-class ParentJoinSecondFragment:BaseFragment<FragmentParentJoinSecondBinding, ParentJoinSecondViewModel>(
-    R.layout.fragment_parent_join_second
-) {
-    override val viewModel: ParentJoinSecondViewModel by viewModels()
-
-    override fun initView() {
-//        TODO("Not yet implemented")
-    }
-}
