--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/second/ParentJoinSecondViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.second
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class ParentJoinSecondViewModel:BaseViewModel() {
-}
