--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/third/ParentJoinThirdFragment.kt
@@ -1,16 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.third
-
-import androidx.fragment.app.viewModels
-import com.b1nd.alimo.R
-import com.b1nd.alimo.base.BaseFragment
-import com.b1nd.alimo.databinding.FragmentParentJoinThirdBinding
-
-class ParentJoinThirdFragment:BaseFragment<FragmentParentJoinThirdBinding, ParentJoinThirdViewModel>(
-    R.layout.fragment_parent_join_third
-) {
-    override val viewModel: ParentJoinThirdViewModel by viewModels()
-
-    override fun initView() {
-//        TODO("Not yet implemented")
-    }
-}
