--- app/src/main/java/com/b1nd/alimo/onboarding/parent/join/third/ParentJoinThirdViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.join.third
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class ParentJoinThirdViewModel:BaseViewModel() {
-}
