--- app/src/main/java/com/b1nd/alimo/onboarding/parent/login/first/ParentLoginFirstViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.login.first
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class ParentLoginFirstViewModel:BaseViewModel() {
-}
