--- app/src/main/java/com/b1nd/alimo/onboarding/parent/pw/first/ParentFindPWFirstFragment.kt
@@ -1,15 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.pw.first
-
-import androidx.fragment.app.viewModels
-import com.b1nd.alimo.R
-import com.b1nd.alimo.base.BaseFragment
-import com.b1nd.alimo.databinding.FragmentParentFindPwFirstBinding
-
-class ParentFindPWFirstFragment:BaseFragment<FragmentParentFindPwFirstBinding, ParentFindPWFirstViewModel>(
-    R.layout.fragment_parent_find_pw_first) {
-    override val viewModel: ParentFindPWFirstViewModel by viewModels()
-
-    override fun initView() {
-//        TODO("Not yet implemented")
-    }
-}
