--- app/src/main/java/com/b1nd/alimo/onboarding/parent/pw/second/ParnetFindPWSecondFragment.kt
@@ -1,16 +1,0 @@
-package com.b1nd.alimo.onboarding.parent.pw.second
-
-import androidx.fragment.app.viewModels
-import com.b1nd.alimo.R
-import com.b1nd.alimo.base.BaseFragment
-import com.b1nd.alimo.databinding.FragmentParentFindPwFirstBinding
-
-class ParnetFindPWSecondFragment:BaseFragment<FragmentParentFindPwFirstBinding, ParentFindPWSecondViewModel>(
-    R.layout.fragment_parent_find_pw_second
-) {
-    override val viewModel: ParentFindPWSecondViewModel by viewModels()
-
-    override fun initView() {
-//        TODO("Not yet implemented")
-    }
-}
