--- app/src/main/java/com/b1nd/alimo/onboarding/second/OnboardingSecondViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.second
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class OnboardingSecondViewModel:BaseViewModel() {
-}
