--- app/src/main/java/com/b1nd/alimo/onboarding/student/first/StudentLoginViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.student.first
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class StudentLoginViewModel:BaseViewModel() {
-}
