--- app/src/main/java/com/b1nd/alimo/onboarding/third/OnboardingThirdViewModel.kt
@@ -1,6 +1,0 @@
-package com.b1nd.alimo.onboarding.third
-
-import com.b1nd.alimo.base.BaseViewModel
-
-class OnboardingThirdViewModel:BaseViewModel() {
-}
